Test Config for each flow
